rename 'widgets' to 'statistics','validators.tsx' to 'validatorlist.tsx'
