+++ packages/ui/src/app/pages/Validators/ValidatorList.tsx
@@ -4,10 +4,10 @@ import { PageHeader } from '@/app/components/PageHeader'
 import { PageLayout } from '@/app/components/PageLayout'
 import { RowGapBlock } from '@/common/components/page/PageContent'
 import { Statistics } from '@/common/components/statistics'
-import { Era } from '@/validators/components/widgets/Era'
-import { Rewards } from '@/validators/components/widgets/Rewards'
-import { Staking } from '@/validators/components/widgets/Staking'
-import { ValidatorsState } from '@/validators/components/widgets/ValidatorsState'
+import { Era } from '@/validators/components/statistics/Era'
+import { Rewards } from '@/validators/components/statistics/Rewards'
+import { Staking } from '@/validators/components/statistics/Staking'
+import { ValidatorsState } from '@/validators/components/statistics/ValidatorsState'
 import { ValidatorsInfo } from '@/validators/modals/ValidatorsInfo'
 
 import { ValidatorsTabs } from './components/ValidatorsTabs'
