+++ packages/ui/src/validators/components/statistics/Rewards.tsx
