+++ packages/ui/src/validators/components/statistics/Staking.tsx
